[refactor] #10 시설물 도메인 객체 리팩토링
- 필드에서 기존에 생성한 enum들을 사용하도록 수정함
- 위치를 facility/domain/ 으로 변경함
